+++ src/tooling/piral-cli-webpack/src/configs/SheetPlugin.ts
@@ -1,21 +1,21 @@
 import InjectPlugin from 'webpack-inject-plugin';
 
-function sheetLoader() {
+function sheetLoader(cssName: string) {
   return () => {
     const lines = [
       `var d=document`,
       `var e=d.createElement("link")`,
       `e.type="text/css"`,
       `e.rel="stylesheet"`,
-      `e.href=__webpack_public_path__ + ${JSON.stringify('main.css')}`,
+      `e.href=__webpack_public_path__ + ${JSON.stringify(cssName)}`,
       `d.head.appendChild(e)`,
     ];
     return lines.join(';');
   };
 }
 
-export default class SheetPlugin {
-  apply(compiler) {
-    new InjectPlugin(sheetLoader()).apply(compiler);
+export default class SheetPlugin extends InjectPlugin {
+  constructor(cssName: string) {
+    super(sheetLoader(cssName));
   }
 }