+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 * Added runtime pilet metadata `basePath`
 * Added support for new debug utils (#397)
 
+## 0.13.6 (August 27, 2021)
+
+* Fixed issue with CSS imports in pilets when using Webpack (#403)
+* Improved documentation (#395)
+* Improved `piral-debug-utils` with new architecture (#397)
+
 ## 0.13.5 (August 17, 2021)
 
 * Fixed explicit declaration via state of `LanguagesPicker` component